Improve specular

+++ color.hpp
@@ -9,9 +9,9 @@ public:
     Color operator*(const double alpha) const;
 
 public:
-    uint8_t RED;
-    uint8_t GREEN;
-    uint8_t BLUE;
+    uint16_t RED;
+    uint16_t GREEN;
+    uint16_t BLUE;
 };
 
 std::ostream& operator<<(std::ostream &out , const Color &color);
